+++ pkg/inotify/event.go
@@ -1,0 +1,3 @@
+package inotify
+
+type Event struct {}
