+++ pkg/inotify/inotify_test.go
@@ -1,146 +1,13 @@
 package inotify
 
 import (
-	"errors"
-	"fmt"
 	"log/slog"
 	"os"
-	"strings"
-	"syscall"
 	"testing"
-	"unsafe"
 
-	"github.com/google/uuid"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
-	"golang.org/x/sys/unix"
 )
-
-type Watch struct {
-	events chan<- Event
-	logger *slog.Logger
-
-	inotify *os.File
-	buf *inotifyBuffer
-	nameByWatchDescriptor map[int] string
-
-	stopServing chan<- struct{}
-}
-
-type inotifyBuffer struct {
-	data []byte
-	offset int
-	numRead int
-}
-
-type Event struct {}
-
-func NewWatch(events chan<- Event, logger *slog.Logger) (Watch, error) {
-	fd, err := unix.InotifyInit()
-	if err != nil {
-		return Watch{}, err
-	}
-
-	stopServing := make(chan struct{})
-
-	w := Watch{
-		events:                events,
-		logger:                logger,
-		inotify:               os.NewFile(uintptr(fd), uuid.NewString()),
-		buf:                   &inotifyBuffer{},
-		nameByWatchDescriptor: make(map[int]string),
-		stopServing:           stopServing,
-	}
-
-	logger.Info("Initialized new inotify instance", "inotify", w.inotify.Name())
-
-	go w.serveEvents(stopServing)
-
-	return w, nil
-}
-
-type inotifyEvent struct {
-	*unix.InotifyEvent
-	name string
-}
-
-func (w *Watch) Stop() {
-	close(w.stopServing)
-	w.inotify.Close()
-}
-
-func (w *Watch) readEventName(event *unix.InotifyEvent) string {
-	if event.Len == 0 {
-		name, ok := w.nameByWatchDescriptor[int(event.Wd)]
-		if !ok {
-			w.logger.Error("Watch descriptor not found", "wd", int(event.Wd))
-			panic("failed to determine file name of event: watch decriptor not found")
-		}
-		return name
-	} else {
-		buf := w.buf
-		name := strings.TrimRight(string(buf.data[buf.offset:buf.offset+int(event.Len)]), "\x00")
-		buf.offset += int(event.Len)
-		return name
-	}
-
-}
-
-func (w *Watch) readEvent() (*inotifyEvent, error)  {
-	buf := w.buf
-	if buf.offset >= buf.numRead {
-		buf.data = make([]byte, 10*(syscall.SizeofInotifyEvent+unix.NAME_MAX+1)) // NOTE: maybe reuse memory here?
-		var err error
-		buf.numRead, err = w.inotify.Read(buf.data)
-		if err != nil {
-			return nil, err
-		}
-		if buf.numRead == 0 {
-			return nil, errors.New("read() from inotify returned 0")
-		}
-		buf.offset = 0
-		w.logger.Debug("Successful read() from inotify", "numRead", buf.numRead, "inotify", w.inotify.Name())
-	}
-
-	var event inotifyEvent
-	event.InotifyEvent = (*unix.InotifyEvent)(unsafe.Pointer(&buf.data[buf.offset]))
-	buf.offset += syscall.SizeofInotifyEvent
-	event.name = w.readEventName(event.InotifyEvent)
-
-	w.logger.Debug("Read event", "wd", event.Wd, "name", event.name, "len", event.Len,
-	"flags", event.Mask)
-
-	return &event, nil
-}
-
-func (w *Watch) serveEvents(stop <-chan struct{}) {
-	defer w.logger.Info("Stop serving events", "inotify", w.inotify.Name())
-	w.logger.Info("Start serving events", "inotify", w.inotify.Name())
-	for {
-		_, err := w.readEvent()
-		if err != nil {
-			w.logger.Warn(fmt.Sprintf("Got error while reading events: %v", err.Error()))
-			close(w.events)
-			return
-		}
-		select {
-		case w.events <- Event{}:
-			w.logger.Debug("Serve some event") // TODO: extended log
-		case <-stop:
-			w.logger.Debug("serveEvents got stop signal")
-		}
-	}
-}
-
-func (w *Watch) Subscribe(path string) error {
-	wd, err := unix.InotifyAddWatch(int(w.inotify.Fd()), path, unix.IN_CREATE|unix.IN_MODIFY|unix.IN_DELETE|unix.IN_MOVED_TO|unix.IN_MOVED_FROM|unix.IN_MOVE)
-	if err != nil {
-		return err
-	}
-	w.nameByWatchDescriptor[wd] = path
-	w.logger.Info("Subscribed to new path", "wd", wd, "path", path)
-	return nil
-}
 
 type InotifyTestSuite struct {
     suite.Suite
@@ -170,7 +37,7 @@ func (s *InotifyTestSuite) TestCanSubscribeToFile() {
 	file, err := os.CreateTemp(s.root, "test")
 	defer os.Remove(file.Name())
 	w, _ := NewWatch(make(chan Event), slog.Default())
-	defer w.Stop()
+	defer w.Close()
 
 	w.Subscribe(file.Name())
 
@@ -181,7 +48,7 @@ func (s *InotifyTestSuite) TestCanSubscribeToDirectory() {
 	file, err := os.CreateTemp(s.root, "test")
 	defer os.Remove(file.Name())
 	w, _ := NewWatch(make(chan Event), slog.Default())
-	defer w.Stop()
+	defer w.Close()
 
 	w.Subscribe(file.Name())
 
@@ -191,7 +58,7 @@ func (s *InotifyTestSuite) TestCanSubscribeToDirectory() {
 func (s *InotifyTestSuite) TestServeFileCreateEvent() {
 	events := make(chan Event)
 	w, _ := NewWatch(events, slog.Default())
-	defer w.Stop()
+	defer w.Close()
 	w.Subscribe(s.root)
 	path := s.root+"/test"
 	file, _ := os.Create(path)
@@ -201,15 +68,6 @@ func (s *InotifyTestSuite) TestServeFileCreateEvent() {
 	event := <-events
 
 	require.Equal(s.T(), Event{}, event)
-}
-
-func (s *InotifyTestSuite) TestGoroutineCleanup() {
-	{
-		events := make(chan Event)
-		w, _ := NewWatch(events, slog.Default())
-		defer w.Stop()
-		w.Subscribe(s.root)
-	}
 }
 
 func TestExampleTestSuite(t *testing.T) {
